finishes get/post methods for responses

+++ RevatureConnect/src/main/java/com/revature/service/ResponseService.java
@@ -15,7 +15,7 @@ public class ResponseService {
     @Autowired
     private ResponseRepository rrepo;
 
-    public List<Response> getResponsesByPostId(Integer postId) {
+    public List<Response> getResponsesByPostId(long postId) {
         return rrepo.getResponsesByPostId(postId); 
     }
